pyproject.toml
[project]
name = "aeroforge"
version = "0.1.0"
description = "AI-guided heavier-than-air vehicle design system with tracked workflow orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "All Rights Reserved"}
authors = [
    {name = "ipanov"}
]
keywords = ["aircraft", "cad", "workflow", "parametric", "design-orchestration", "build123d"]
dependencies = [
    "build123d>=0.10.0",
    "networkx>=3.2",
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "matplotlib>=3.8.0",
    "pyvista>=0.43.0",
    "meshio>=5.3.0",
    "rich>=13.7.0",
    "ezdxf>=1.3.0",
    "neuralfoil>=0.3.0",
    "trimesh>=4.0.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
